requirements.txt
tushare>=1.3.8
pandas>=2.0.0
numpy>=1.21.0
python-dotenv>=0.19.0
loguru>=0.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
openpyxl>=3.0.0
requests>=2.28.0
tqdm>=4.64.0
# SQLite3已内置在Python中，无需额外安装 